backend/requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.1
django-filter==23.5

# API Documentation
drf-spectacular==0.27.0

# Document Generation
python-docx==1.1.0
openpyxl==3.1.2
Pillow==10.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
django-environ==0.11.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0
